bug fix in type declaration for char f in va_arg

+++ 0x10-variadic_functions/3-print_all.c
@@ -24,7 +24,7 @@ void print_all(const char * const format, ...)
 				printf("%s%d", separator, va_arg(args, int));
 				break;
 			case 'f':
-				printf("%s%f", separator, va_arg(args, float));
+				printf("%s%f", separator, va_arg(args, double));
 				break;
 			case 's':
 				str =  va_arg(args, char *);
